qcommit: Dont check if the repository is locked before saving the commit message.
I'm not sure why this was previously done. I also think we may be able to remove this check completly, as we now have better error handeling, and so if the branch is locked, they will get a error, but they will be able to click "ignore" and the dialog will close anyway.